Need to overload properties to fix api discrepencies
Finding small issues like the S3 API uses a lot of Keys of
Bucket but in actuallity the call is bucketName. We can easily
convert all our cases to Pascal, and even accept overloads
of arbitrary key/value, but at some point I want to coerce
the string key into the proper PascalType.

I can't do this through inference alone since there is no
reflection. I'm already generating the types for the API so
I might as well handle the registry and overloading there to
manipulate the documented API into actual logs in cloudtrail.

Current idea is to:

 - Generate all the interfaces using EJS
 - Handle all the mapping in the generate method and keep the template
clean
 - During map generation read the overloads
 - Apply the overloads to the API to reflect actual values
 - Overloads go into the scripts/overload package
 - Create a registry map for overloads to automatically make them
available
 - Apply overloads from the map as stated above